requirements_new.txt
# Core Flask dependencies
flask==2.3.3
flask-cors==4.0.0
flask-mail==0.9.1
bcrypt==4.0.1

# Configuration and environment
python-dotenv==1.0.0

# Database and caching
redis==5.0.1

# Development and testing
pytest==7.4.3
pytest-flask==1.2.0
pytest-mock==3.12.0

# Additional utilities
werkzeug==2.3.7
jinja2==3.1.2
markupsafe==2.1.3
itsdangerous==2.1.2
blinker==1.7.0

# Security and validation
WTForms==3.0.1
email-validator==2.1.0

# Performance and monitoring
gunicorn==21.2.0
